Fleshed out specator and other backends.
