+++ src/lib.rs
@@ -1,7 +1,10 @@
 pub mod rbackend;
 pub mod revent;
 pub mod rmercury;
+pub mod rpeer2peer;
 pub mod rplayer;
+pub mod rspectator;
+pub mod rsynctest;
 pub mod rsystem;
 #[cfg(test)]
 mod tests {
